modified view helpers further and fixed tests
