+++ spec/features/create_post_spec.rb
