+++ spec/features/user_logs_in_spec.rb
