+++ spec/features/visitor_signs_up_spec.rb
