--- tmp/pids/server.pid
@@ -1,1 +1,0 @@
-2538
